Move common lesson models to lessons-data module
