+++ buildSrc/src/main/kotlin/BuildModules.kt
@@ -3,6 +3,7 @@ object BuildModules {
     object Common {
         const val CORE = ":common:core"
         const val DESIGN = ":common:design"
+        const val LESSONS_DATA = ":common:lessons-data"
         const val TRANSLATIONS = ":common:translations"
     }
