+++ common/lessons-data/src/main/java/app/ss/lessons/data/model/SSDay.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020 Adventech <info@adventech.io>
+ * Copyright (c) 2016 Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -19,33 +19,22 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
+package app.ss.lessons.data.model
 
-package com.cryart.sabbathschool.lessons.data.model.response
+import androidx.annotation.Keep
+import com.google.firebase.database.IgnoreExtraProperties
 
-import com.cryart.sabbathschool.lessons.data.model.response.Status.ERROR
-import com.cryart.sabbathschool.lessons.data.model.response.Status.LOADING
-import com.cryart.sabbathschool.lessons.data.model.response.Status.SUCCESS
-
-class Resource<out T> private constructor(
-    val status: Status = LOADING,
-    val data: T?,
-    val error: Throwable?
+@Keep
+@IgnoreExtraProperties
+data class SSDay(
+    val title: String,
+    val date: String = "",
+    val id: String = "",
+    val index: String = "",
+    val path: String = "",
+    val full_path: String = "",
+    val read_path: String = "",
+    val full_read_path: String = "",
 ) {
-
-    val isSuccessFul: Boolean get() = data != null
-
-    companion object {
-
-        fun <T> success(data: T): Resource<T> {
-            return Resource(SUCCESS, data, null)
-        }
-
-        fun <T> error(error: Throwable): Resource<T> {
-            return Resource(ERROR, null, error)
-        }
-
-        fun <T> loading(): Resource<T> {
-            return Resource(LOADING, null, null)
-        }
-    }
+    constructor() : this("")
 }
