+++ common/lessons-data/src/main/java/app/ss/lessons/data/model/SSLesson.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020 Adventech <info@adventech.io>
+ * Copyright (c) 2016 Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -19,29 +19,22 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.cryart.sabbathschool.lessons.data.model
+package app.ss.lessons.data.model
 
 import androidx.annotation.Keep
 import com.google.firebase.database.IgnoreExtraProperties
 
 @Keep
 @IgnoreExtraProperties
-class SSQuarterly(
-    val id: String,
-    val title: String = "",
-    val description: String = "",
-    val human_date: String = "",
+class SSLesson(
+    val title: String,
     val start_date: String = "",
     val end_date: String = "",
     val cover: String = "",
+    val id: String = "",
     val index: String = "",
-    val group: String? = null,
     val path: String = "",
     val full_path: String = "",
-    val lang: String = "",
-    val color_primary: String,
-    val color_primary_dark: String = "",
-    val quarterly_name: String = "",
 ) {
-    constructor() : this("", color_primary = "")
+    constructor() : this("")
 }
