+++ common/lessons-data/src/main/java/app/ss/lessons/data/model/SSReadComments.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016 Adventech <info@adventech.io>
+ * Copyright (c) 2021. Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -19,21 +19,22 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.cryart.sabbathschool.lessons.data.model
+package app.ss.lessons.data.model
 
 import androidx.annotation.Keep
 import com.google.firebase.database.DataSnapshot
+import com.google.firebase.database.IgnoreExtraProperties
 
 @Keep
-data class SSQuarterlyInfo(
-    val quarterly: SSQuarterly,
-    val lessons: List<SSLesson>
+@IgnoreExtraProperties
+data class SSReadComments(
+    val readIndex: String,
+    val comments: List<SSComment>
 ) {
-
-    constructor(snapshot: DataSnapshot) : this(
-        snapshot.child("quarterly").getValue(SSQuarterly::class.java) ?: SSQuarterly(),
-        snapshot.child("lessons").children.mapNotNull {
-            it.getValue(SSLesson::class.java)
+    constructor(snapshot: DataSnapshot, dayIndex: String) : this(
+        snapshot.child("readIndex").getValue(String::class.java) ?: dayIndex,
+        snapshot.child("comments").children.mapNotNull {
+            it.getValue(SSComment::class.java)
         }
     )
 }
