+++ common/lessons-data/src/main/java/app/ss/lessons/data/model/SSSuggestion.kt
@@ -19,22 +19,17 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.cryart.sabbathschool.lessons.data.model
+package app.ss.lessons.data.model
 
 import androidx.annotation.Keep
-import com.google.firebase.database.DataSnapshot
 import com.google.firebase.database.IgnoreExtraProperties
 
 @Keep
 @IgnoreExtraProperties
-data class SSLessonInfo(
-    val lesson: SSLesson,
-    val days: List<SSDay>
+data class SSSuggestion(
+    val name: String,
+    val email: String = "",
+    val suggestion: String = "",
 ) {
-    constructor(snapshot: DataSnapshot) : this(
-        snapshot.child("lesson").getValue(SSLesson::class.java) ?: SSLesson(),
-        snapshot.child("days").children.mapNotNull {
-            it.getValue(SSDay::class.java)
-        }
-    )
+    constructor() : this("")
 }
