+++ common/lessons-data/src/main/java/app/ss/lessons/data/response/Status.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016 Adventech <info@adventech.io>
+ * Copyright (c) 2020 Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -19,22 +19,11 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.cryart.sabbathschool.lessons.data.model
 
-import androidx.annotation.Keep
-import com.google.firebase.database.IgnoreExtraProperties
+package app.ss.lessons.data.response
 
-@Keep
-@IgnoreExtraProperties
-data class SSDay(
-    val title: String,
-    val date: String = "",
-    val id: String = "",
-    val index: String = "",
-    val path: String = "",
-    val full_path: String = "",
-    val read_path: String = "",
-    val full_read_path: String = "",
-) {
-    constructor() : this("")
+enum class Status {
+    LOADING,
+    SUCCESS,
+    ERROR
 }
