+++ features/lessons/src/main/java/com/cryart/sabbathschool/lessons/ui/quarterlies/QuarterliesActivity.kt
@@ -38,7 +38,7 @@ import com.cryart.sabbathschool.core.model.ViewState
 import com.cryart.sabbathschool.core.navigation.AppNavigator
 import com.cryart.sabbathschool.core.navigation.Destination
 import com.cryart.sabbathschool.lessons.R
-import com.cryart.sabbathschool.lessons.data.model.SSQuarterly
+import app.ss.lessons.data.model.SSQuarterly
 import com.cryart.sabbathschool.lessons.databinding.SsActivityQuarterliesBinding
 import com.cryart.sabbathschool.lessons.ui.base.SSBaseActivity
 import com.cryart.sabbathschool.lessons.ui.languages.LanguagesListFragment
